requirements-dev.txt
# Development requirements
pyproj
matplotlib
cartopy
pytest
pytest-cov
coverage
pylint
flake8
sphinx=2.2.1
sphinx_rtd_theme=0.4.3
sphinx-gallery
numpydoc
twine
